Add alignment prop
This is necessary to work property when the Resentence container changes
size based on the length of the text- we need to know the "fixed" point
to correctly compute the positions when animating the transition. This
can't easily be determined as it can depend on things such as flex
containers around the Resentence element, so we require the consumer to
pass in the alignment as a prop.